Clearer failure messages
Previously, the following failing assertion:

    expect(true).not_to be_boolean

...would result in this failure message:

    expected that true would be a true or false

This wording sounds like we were expecting the value to be `true`, but
our spec will only pass if it's _not_ `true` or `false`.

With this change, the error message for negative assertions such as
`not_to` will more clearly explain what the spec was expecting.